+++ src/pages/applications/Applicantions.css
@@ -76,15 +76,7 @@ body {
   padding-left: 20px;
 }
 
-<<<<<<< HEAD:src/pages/adminDashoard/css/index.css
-.menu-container {
-  margin-top: 30px;
-=======
-.menu-container{
-  margin: 26px 0px;
-  padding: 10px;
->>>>>>> 7b6c21022e61d3e112500aaf189555ed9fc8acd5:src/pages/applications/Applicantions.css
-}
+
 
 .logo {
   padding: 47px 40px 0px;
@@ -100,11 +92,7 @@ body {
   padding: 20px;
 }
 
-<<<<<<< HEAD:src/pages/adminDashoard/css/index.css
 .search__bar {
-=======
-.header-controls .search__bar{
->>>>>>> 7b6c21022e61d3e112500aaf189555ed9fc8acd5:src/pages/applications/Applicantions.css
   max-width: 200px;
   width: 100%;
 }
